+++ app/src/main/java/k0bin/notes/ui/viewHolder/NoteViewHolder.java
@@ -14,7 +14,7 @@ import k0bin.notes.viewModel.NotesViewModel;
 public class NoteViewHolder extends RecyclerView.ViewHolder {
 	private final TextView title;
 	private final TextView text;
-	private int noteId = 0;
+	private long noteId = 0;
 
 	public NoteViewHolder(@NonNull View itemView) {
 		super(itemView);
@@ -23,7 +23,7 @@ public class NoteViewHolder extends RecyclerView.ViewHolder {
 		this.text = itemView.findViewById(R.id.text);
 
 		itemView.setOnClickListener(v -> {
-			EditFragmentArgs args = new EditFragmentArgs.Builder().setNoteId(noteId).build();
+			EditFragmentArgs args = new EditFragmentArgs.Builder().setNoteId((int)noteId).build();
 			Navigation.findNavController(v).navigate(R.id.action_notesFragment_to_editFragment2, args.toBundle());
 		});
 	}
